msm8937_32: Start mdtpd (MDTP daemon)
MDTP daemon was added. The daemon replaces the existing
MDTP service.

The mdtpd needs to run as root user, as it requires access
to the system, mdtp and dip block devices, under /dev/block.
The mdtpd doesn't need the root group.
After starting the threads requiring root access, the mdtpd

Change-Id: If5f4d3c8b053d325375beb76e3e14aaf5dd37bc6